Add --with-jpeg-from-wx configure option
This (obviously wx-specific) option short-circuits libjpeg detection and
relies on wxWidgets build of this library being used.

This is more than just an optimization, as using JPEG compression in the
built-in version of the library when also using built-in libjpeg only
worked before if we had a compatible version of libjpeg globally
available: otherwise, libtiff configure either didn't detect libjpeg at
all or, even worse, detected an incompatible version different from the
one actually used when building.